+++ src/modules/logs/content/log_event_content/scientific_research_event.rs
@@ -8,5 +8,5 @@ pub struct ScientificResearchEvent {
     pub market_id: u64,
     pub name: Material,
     pub category: MaterialCategory,
-    pub count: u8,
+    pub count: u16,
 }
